+++ client/src/pages/login/login.css
@@ -1,0 +1,7 @@
+.logincontainer {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
